+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Flutter plugin for selecting files from the Android and iOS file library, and taking new pictures with the camera.